% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the statline app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[shiny]{shinyApp}} in the
Shiny package.}
}
\value{
Nothing
}
\description{
Run the statline app
}
\examples{
\dontrun{
run_app()
}

}
